requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.45
langchain-core==0.3.15
langchain-openai==0.2.8
langsmith==0.1.147
httpx==0.27.2
python-dotenv==1.0.0
pytest==8.3.3
pytest-asyncio==0.24.0